feat(pkgs/database): prep for prisma accelerate
