+++ packages/database/package.json
@@ -14,9 +14,26 @@
   "license": "MIT",
   "author": "Matthew Lilley <hello@matthewLilley.com>",
   "type": "module",
-  "main": "dist/index.js",
-  "source": "src/index.ts",
-  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/node.d.ts",
+      "default": "./dist/node.js"
+    },
+    "./edge": {
+      "types": "./dist/edge.d.ts",
+      "default": "./dist/edge.js"
+    }
+  },
+  "main": "dist/node.js",
+  "source": "src/node.ts",
+  "types": "dist/node.d.ts",
+  "typesVersions": {
+    "*": {
+      "edge": [
+        "./dist/edge.d.ts"
+      ]
+    }
+  },
   "files": [
     "dist"
   ],
@@ -36,7 +53,8 @@
     "preset": "@sushiswap/jest-config"
   },
   "dependencies": {
-    "@prisma/client": "4.12.0"
+    "@prisma/client": "4.16.2",
+    "@prisma/extension-accelerate": "^0.6.2"
   },
   "devDependencies": {
     "@sushiswap/jest-config": "workspace:*",
@@ -46,7 +64,7 @@
     "@types/node": "18",
     "dotenv": "16.3.1",
     "jest": "29.7.0",
-    "prisma": "4.12.0",
+    "prisma": "4.16.2",
     "typescript": "5.2.2"
   }
 }
