--- packages/database/tsup.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from 'tsup'
-
-export default defineConfig({
-  minify: true,
-  target: 'esnext',
-  external: [],
-  sourcemap: true,
-  dts: true,
-  format: ['esm'],
-})
